fix : distance data type to double
Former-commit-id: 34b0fc76b9a51c424772daad631ed42c2d20bdd2
Former-commit-id: 3dfd5089ac87167009dc85d661e4fd13e0ea25a6

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/food/dto/request/FoodRequestDto.java
@@ -12,7 +12,7 @@ public class FoodRequestDto {
 	@AllArgsConstructor
 	public static class Title {
 		private String attractionName;
-		private Long distance;
+		private Double distance;
 		private String sorted;
 	}
 
@@ -23,7 +23,7 @@ public class FoodRequestDto {
 	public static class Coordi {
 		private Double latitude;
 		private Double longitude;
-		private Long distance;
+		private Double distance;
 		private String sorted;
 	}
 }
